+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 - Support generating a flamegraph from a list of chunk IDs ([#476](https://github.com/getsentry/vroom/pull/476))
 - Dual mode flamegraph endpoint ([#487](https://github.com/getsentry/vroom/pull/487))
 - Remove getProfileIDFromTransaction endpoint ([#491](https://github.com/getsentry/vroom/pull/491))
+- Pass thread id to calltree generation ([#492](https://github.com/getsentry/vroom/pull/492))
+- Dual mode metrics endpoint ([#493](https://github.com/getsentry/vroom/pull/493))
 
 **Bug Fixes**: